admin: update - Add new validators and form components
Adds the form component 'TextInput' and replaces the input on the login
page.

Adds the form component 'PasswordInput' and replaces the input on the
login page.

Adds the form component 'HiddenInput'

Adds the form component 'SubmitButton' and replaces the button on the
login page.

Defines a lot of mixins that are used be components mentioned above.

Adds the following validators:
  - autocompleteValidator
  - iconValidator
  - inputmodeValidator

Adds tests for each component and validator

Updates the LoginPage.spec.js

Adds some new locales

Signed-off-by: Sebastian Krah <exotelis@mailbox.org>