admin: update - Improve the NavigationGenerator and NavigationItem
This commit improves the NavigationGenerator and the NavigationItem
components. Instead of the wrapper workaround we used before, we
switched to protal-vue instead. This change has been made to clean up
the scss files and in preparation of vue 3 portals. Adds documentation
of all the improvements. Different styles of navigation can be
generated. No matter if horizontal oder vertically. The name of the
active class, which the router sets, has been changed to active.