api: update - Update clear users job
The reason why it is configurable for how long users should be stored,
is that you have to keep bills etc. for a longer period. In some
countries those periods begin at the end of the current year. So in some
cases you have to store customer data for 10 years and some more months.
To gurantee that those data won't be deleted too early, we just remove
them at the end of the year. However the job will still run every day to
make sure that users with unverified email will still be deleted.