api: update - Improve event handling and clean up
This commit is more or less a general clean up. It replaces all events
that have been called manually with observers. Observers listen to
models events like saving, updating, deletes or restoring. Whenever
such an operation is executed the suitable method of the observer is
called. Apart from this users can restore their account when they have
been soft deleted now. This is helpful when a user was inactive for a
longer period and the account is marked for deletion.
The log messages have been restructured. The type of the message is
printed in square brackets at the beginning of each message.

- Adds resource auth/recover, that a user can recover a soft deleted
  account with a token. Once a account gets soft deleted the user can
  recover the account for a specific period.
- Adds resource users/:id/restore to restore a user.
- Adds and updates tests.
- Adds restore and soft delete notifications, mails and templates.
- Adds restore_users migration.
- Adds helper functions getModelDiff and recoverUsers.
- Adds job to clean up the restore_users table.
- Adds the RestoreRequest to validate the requests on resource
  auth/recover.
- Adds trait ExceptRoute which defines the except array that is used by
  many middleware.
- General cleanup.
- Updates the log messages to give them a more structured pattern.
- Updates l10n.
- Updates api documentation
- Removes a lot of events that have been triggered when a user has been
  created, deleted etc. We moved to observers to handle those events in
  the same place.