api: update - User deletion management
This commit adds two new jobs. The DeleteInactiveUsers job, is going to
delete users that have been inactive for a given period of time. The job
is checking the updated_at field to determine if an user is active or
not. Inactive users will be soft deleted. Once a user as been soft
deleted the ClearUsers job is going to deleted all users that haven't
recovered their accounts after they have been marked as deleted.

The DeleteUnverifiedUsers job did exist already, but was part of the
ClearUsers job. It has been moved out for a better separation.

Fixes: https://github.com/Exotelis/skydivemanifest/issues/27