api: add - Qualifications management
Adds qualifications management to the api. Each user can have none, one
or multiple qualifications, such as packer or load organizer. This might
be useful, when the dropzone needs to plan the jumping days. Later, a
user can book himself for a specific day. This will help the organizer
to plan if enough staff is available.

- Adds Qualification migration.
- Adds QualificationUser migration.
- Adds Qualification model.
- Adds Qualification factory and seeder.
- Adds /qualifications routes.
- Adds QualificationController as well as tests.
- Adds Create, DeleteBulk and Update requests for the qualification
  controller.
- Adds QualificationFilters.
- Adds QualificationObserver.
- Qualifications of users can be changed, and any user can change own
  qualifications. Added all required routes, methods and tests.
- Updates the User model and adds the qualifications relation.
- Updates l10n.