api: add - Waiver management
Adds waiver management to the api. Waivers can be created, updated or
deleted. Each waiver kann have multiple texts, those can also get
repositioned. Waiver can not yet be assigned to a user, this will be
possible with the next commit. Furthermore, it is possible to request
all active waivers.

- Adds Waiver and Text migrations.
- Adds Waiver and Text models.
- Adds Waiver and Text factories.
- Adds Waiver seeder.
- Adds /waivers routes.
- Adds Waiver and Text controllers as well as tests.
- Adds create and update requests for both the Waiver and the Text
  controllers.
- Adds Waiver and Text filters.
- Adds Waiver and Text observers.
- Updates user requests and added custom attribute text for is_active.
- Updates the DemoSeeder.
- Updates l10n.